Optimization using real or transformed parameters
Implemented parameter optimization using both real or transformed parameters.
For now, it looks like it's best to use real parameter for optimization.
However, results are not as good as the ones obtained by the excell solver
closes #5

Included limits to the parameter estimation in the optimization functions.
Uses the bounds parameter
closes #8